requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
coverage==7.6.12
Django==5.1.6
idna==3.10
iniconfig==2.0.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg2==2.9.10
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
